+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/runner/NativeExecCommand.scala
@@ -21,9 +21,6 @@ case class NativeExecCommand(executablePath: Path) extends ExecCommand {
 }
 
 object NativeExecCommand {
-
-  private val LAUNCHER_ENV_NAME = "ENSO_LAUNCHER"
-
   def apply(
     version: String,
     engine: Engine,
@@ -34,11 +31,10 @@ object NativeExecCommand {
     val execName = OS.executableName("enso")
     val fullExecPath =
       dm.paths.engines.resolve(version).resolve("bin").resolve(execName)
-    val ensoLauncher = Option(System.getenv(LAUNCHER_ENV_NAME))
 
     if (fullExecPath.toFile.exists() && isBinary(fullExecPath, logger)) {
       Some(NativeExecCommand(fullExecPath))
-    } else if (ensoLauncher.map(_.equals("native")).getOrElse(false)) {
+    } else {
       val component =
         engine.componentDirPath.resolve("..").toAbsolutePath.normalize
       val fallbackExecPath = component.resolve("bin").resolve(execName)
@@ -47,14 +43,12 @@ object NativeExecCommand {
       ) {
         Some(NativeExecCommand(fallbackExecPath))
       } else {
-        logger.debug(
+        logger.warn(
           "Failed to find native launcher at a pre-determined location: {}",
           fallbackExecPath
         )
         None
       }
-    } else {
-      None
     }
   }