+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - [Warning and Error messages does not obstruct output port][12482] - they
   become semi-transparent on port hover and pass all mouse interactions.
 - [For some types, Component Browser display special "suggestions" group][12477]
+- [Native Image mode is now on by default][12515]
 
 [11889]: https://github.com/enso-org/enso/pull/11889
 [11836]: https://github.com/enso-org/enso/pull/11836
@@ -75,6 +76,7 @@
 [12496]: https://github.com/enso-org/enso/pull/12496
 [12482]: https://github.com/enso-org/enso/pull/12482
 [12477]: https://github.com/enso-org/enso/pull/12477
+[12515]: https://github.com/enso-org/enso/pull/12515
 
 #### Enso Standard Library
 
@@ -138,6 +140,8 @@
 - [Use fn... to reference any module function][12128]
 - [Improve error message for mismatched named argument application][12238]
 - [Registering a value as multiple managed resources is now an error.][12395]
+- [An operator block now applies to the whole preceding expression][12505],
+  rather than the last term on the line.
 
 [11777]: https://github.com/enso-org/enso/pull/11777
 [11600]: https://github.com/enso-org/enso/pull/11600
@@ -151,6 +155,7 @@
 [12128]: https://github.com/enso-org/enso/pull/12128
 [12238]: https://github.com/enso-org/enso/pull/12238
 [12395]: https://github.com/enso-org/enso/pull/12395
+[12505]: https://github.com/enso-org/enso/pull/12505
 
 # Enso 2024.5